Patched main.py to correctly use ESP charges when making .mfj inputs
In the previous patch, a crucial segment of code responsible for appending ESP (Electrostatic Potential) data to the .key file, generated by the sdf2tinkerxyz package, was unintentionally deleted. As a consequence, the .key files only captured charge information related to MMFF94, neglecting the charges derived from ESP mapping. This oversight led to erroneous CCS calculations due to the presence of inaccurate partial charges in the .mfj input file.

The issue has been rectified, and the GUI now correctly generates .mfj input files. We sincerely apologize for any inconvenience caused by this oversight.

In this patch, we define the variable for file_num that was not being indexed correctly. Don't debug before your morning coffee, friends!